requirements.txt
absl-py==2.3.0
alabaster==1.0.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
box2d-py==2.3.5
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
comm==0.2.2
commonmark==0.9.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.4
docutils==0.21.2
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
gymnasium==1.1.1
h5py==3.14.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.66.2
idna==3.10
imageio==2.37.0
imagesize==1.4.1
ipykernel==6.29.5
ipython==8.37.0
jax==0.4.25
jaxlib==0.4.25+cuda11.cudnn86
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
loguru==0.7.3
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
msgpack-numpy==0.4.8
multimethod==2.0
multiprocess==0.70.12.2
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbsphinx==0.9.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathos==0.2.8
pbr==6.1.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
ply==3.11
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
pyscf==2.9.0
pyspnego==0.11.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qiskit==0.46.3
qiskit-aer==0.13.3
qiskit-ibm-runtime==0.20.0
qiskit-terra==0.46.3
qutip==5.2.0
recommonmark==0.7.1
referencing==0.36.2
requests==2.32.4
requests_ntlm==1.3.0
rich==14.0.0
rpds-py==0.25.1
rustworkx==0.16.0
scipy==1.15.3
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
stevedore==5.4.1
swig==4.3.1
symengine==0.14.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
tensorpack==0.11
termcolor==3.1.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.7.1
torchdiffeq==0.2.5
torchpack==0.3.1
-e git+https://github.com/mit-han-lab/torchquantum.git@00039cc9cd2368f9063b6535eefd4981183a84ea#egg=torchquantum
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
